requirements.txt
scikit-image
Pillow 
imageio
